% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docket_search.R
\name{docket_search}
\alias{docket_search}
\title{SCOTUS Docket Search}
\usage{
docket_search(
  docket_id,
  rate = 50,
  sleep = 30,
  include = NULL,
  exclude = NULL,
  write_csv = FALSE,
  write_xlsx = FALSE,
  write_reactable = FALSE
)
}
\arguments{
\item{docket_id}{Character or comma-separated vector of docket ID numbers (e.g., "19-1392")}

\item{rate}{Rate of docket processing (default is **50**)}

\item{sleep}{Seconds to wait between docket requests (default is **30**)}

\item{include}{Character vector to limit the output variables to only those requested (default is **NULL**).}

\item{exclude}{Character vector that will exclude any output variables requested (default is **NULL**)}

\item{write_csv}{Logical value indicating whether to print code to convert the output dataframe to a .csv file (default is **FALSE**)}

\item{write_xlsx}{Logical value indicating whether to print code to convert the output dataframe to a excel file (default is **FALSE**)}

\item{write_reactable}{Logical value indicating whether to print code to convert the output dataframe to a .html file using `reactable()` (default is **FALSE**).}
}
\value{
Parsed Supreme Court Docket Information
}
\description{
SCOTUS Docket Search
}
\examples{
docket_search(docket_id = "19-1392")
}
