% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa_parser.R
\name{oa_parser}
\alias{oa_parser}
\title{SCOTUS Oral Argument Parser}
\usage{
oa_parser(dir_path = NULL)
}
\arguments{
\item{dir_path}{Folder directory path to a folder on your local machine containing oral argument transcripts from the United States Supreme Court. *Note*: Files must be in PDF format.}
}
\value{
Parsed and cleaned text of oral argument transcripts, separated by speaker, text/statement, term, speaker type (i.e., Justice or Attorney), argument (case title), argument docket number, and word count.
}
\description{
SCOTUS Oral Argument Parser
}
\examples{
\dontrun{
  #Replace "<Folder Directory>" with appropriate location.
  sample_oral_argument <- oa_parser(dir_path = "<Folder Directory>")
}
}
