% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oa_search.R
\name{oa_search}
\alias{oa_search}
\title{SCOTUS Oral Argument Search}
\usage{
oa_search(
  term = NULL,
  justice = NULL,
  attorney = NULL,
  speaker_type = NULL,
  docket_id = NULL,
  party = NULL
)
}
\arguments{
\item{term}{Term of interest as character or vector}

\item{justice}{Justice(s) of interest as character or vector}

\item{attorney}{Attorney(s) of interest as character or vector}

\item{speaker_type}{Factor indicating return Justices or Attorneys only}

\item{docket_id}{Supreme Court docket identification number coinciding with argument(s) of interest}

\item{party}{Party or parties represented in argument(s) of interest}
}
\value{
Parsed data frame of oral argument data (2004 to present or reflective of optional parameters)
}
\description{
SCOTUS Oral Argument Search
}
\examples{
sample_oa <- oa_search(term = "2005", justice = "Breyer")
}
