% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oyez_argument_scraper.R
\name{oyez_transcript_search}
\alias{oyez_transcript_search}
\title{Oyez Oral Argument Transcript Retrieval & Processor}
\usage{
oyez_transcript_search(docket = NULL, term = NULL)
}
\arguments{
\item{docket}{Character denoting Supreme Court docket number (e.g., '19-1392')}

\item{term}{Character denoting Supreme Court term (e.g., '2021')}
}
\value{
Returns dataframe object with cleaned and parsed oral argument transcript
}
\description{
Oyez Oral Argument Transcript Retrieval & Processor
}
\examples{
\dontrun{
dobbs_sample <- oyez_transcript_search(docket = '19-1392', term = '2021')
}
}
