% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_flattened_list.R
\name{print_flattened_list}
\alias{print_flattened_list}
\title{Print Flattened List (scotustext)}
\usage{
print_flattened_list(column)
}
\arguments{
\item{column}{Dataframe and index of interest to convert data structured as flattend semi-colon separated list object ("; ") to nested list format.}
}
\value{
Nested list format of flattend data.
}
\description{
Print Flattened List (scotustext)
}
\examples{
\dontrun{
dobbs_sample <- docket_search(docket_id = "19-1392")
print_flattened_list(dobbs_sample$all_other_counsel)
}
}
