% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docket_search_v2.R
\name{docket_search}
\alias{docket_search}
\title{SCOTUS Docket Search (2.0.0)}
\usage{
docket_search(docket_id)
}
\arguments{
\item{docket_id}{The docket number of interest - written as a character vector in R. This can take the form of filed petitions (e.g., “19-1392”), applications (e.g., “22A948”), or motions (e.g., “22M11”). docket_id parameter is REQUIRED for all searches.}
}
\value{
Cleaned and parsed information from docketed petitions, applications, and motions to the United States Supreme Court.
}
\description{
SCOTUS Docket Search (2.0.0)
}
\examples{
\dontrun{
sample_docket <- docket_search(docket_id = '19-1392')
}
}
