% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docket_search.R
\name{docket_search}
\alias{docket_search}
\title{SCOTUS Docket Search}
\usage{
docket_search(
  docket_id,
  rate = 5000,
  sleep = 30,
  include = NULL,
  exclude = NULL,
  write_csv = FALSE,
  write_xlsx = FALSE,
  write_reactable = FALSE
)
}
\arguments{
\item{docket_id}{The docket number of interest - written as a character vector in R. This can take the form of filed petitions (e.g., “19-1392”), applications (e.g., “22A948”), or motions (e.g., “22M11”). docket_id parameter is REQUIRED for all searches. *Note*: The function will also accept list objects containing comma-separated character vectors of docket numbers (e.g., docket_ids <- c(“19-108”, “19-184”, “19-199”, “19-251”), as well as original jurisdiction petitions (e.g., "22O145")}

\item{rate}{The number of requests to process per function call (default is 5000).}

\item{sleep}{The number of seconds to wait between docket requests (default is 30 seconds).}

\item{include}{A character vector that will limit the output variables to only those requested (default is NULL).}

\item{exclude}{A character vector that will exclude any output variables requested (default is NULL)}

\item{write_csv}{Logical value indicating whether to print code to convert the output dataframe to a .csv file (default is FALSE)}

\item{write_xlsx}{Logical value indicating whether to print code to convert the output dataframe to an excel file (default is FALSE)}

\item{write_reactable}{Logical value indicating whether to print code to convert the output dataframe to a .html file using reactable() (default is FALSE)}
}
\value{
Cleaned and parsed information from docketed petitions, applications, and motions to the United States Supreme Court.
}
\description{
SCOTUS Docket Search
}
\examples{
\dontrun{
sample_docket <- docket_search(docket_id = c("19-1392", "22A122", "22M101", "22O145"))
}
}
